SignUp: Async Validate Issue Fix
Checking for empty object instead of null object
Signup form was not working due to improper merge happened in recent
past.
All changes are retrofitted and asyncValidate has been rectified to
avoid
uncaught error in promise. This was due to the fcat that payload was
returining
error={} which was succssful always for if(error) check. It has been
amended to
look for proper error.msg to throw error for existing user.

Signed-off-by: Progyan Bhattacharya <progyanb@acm.org>