Chnages to refresh the data grid with selected columns for viewing
Added a variable 'selectedViewColumns' for passing to RegOpzFlatGrid columns
to refresh the data grid. Original list of columns are passed to ShowToggleColumns
to select the list of columns.
The change is also propagated for view data as well.
//TODO
At present selected list of columns are not retained when ShowToggleColumns is called
ShowToggleColumns ideally be part of the commom component so that it can be refered by different components.
The data grid displays properly if we select 3 or more columns, however braeks for less than 3 columns!!
Require further testing of other functionalities e.g. insert, duplicate, edit and delete
Export csv for reduced set of columns still exports all the data. Should we prompt users to select the option all data vs selected columns? Suggestions welcome...